software/apps/analytics/requirements.txt
setuptools==45.0.0
numpy==1.16.0
tensorflow-gpu==1.13.1
scipy==1.2.1
scikit-learn==0.21.2
h5py==2.9.0
matplotlib==3.0.3
Pillow==6.1.0
requests==2.2.0
psutil==5.6.3
keras==2.2.4
scikit-image==0.15.0
imutils==0.5.2
pyyaml==5.3.1
confuse==1.1.0
psycopg2==2.8.3